Implement undo functionality, remove randomization

+++ img-mm.py
@@ -24,8 +24,12 @@ app = flask.Flask(__name__)
 APP_URL = 'http://127.0.0.1:5000/'
 RANK_MULTIPLIER = 1000
 TEMPLATE = "img-mm.tpl"
-MU_XATTR = "ts.mu"
-SIGMA_XATTR = "ts.sigma"
+MU_XATTR = "img-mm.ts.mu"
+SIGMA_XATTR = "img-mm.ts.sigma"
+PREV_MU_XATTR = "img-mm.ts.mu.prev"
+PREV_SIGMA_XATTR = "img-mm.ts.sigma.prev"
+PREV_INDEX_XATTR = "img-mm.index.prev"
+PREV_FILENAME_XATTR = "img-mm.filename.prev"
 SUPPORTED_EXTS = [
     ".bmp"
     ".gif",
@@ -37,6 +41,8 @@ SUPPORTED_EXTS = [
 
 eligible_files = []
 
+files_index = {}
+
 def rank(rating):
     rank = str(
         (50 * RANK_MULTIPLIER) - round(
@@ -45,35 +51,52 @@ def rank(rating):
     )
     return rank
 
+def get_xattr(filename, name):
+    try:
+        value = xattr.getxattr(filename, name).decode("UTF8")
+    except OSError:
+        value = None
+    return value
+
+def set_xattr(filename, name, value):
+    xattr.setxattr(filename, name, value.encode('UTF8'))
+
 def get_file(filename):
-    file_xattr = xattr.xattr(filename)
-    try:
-        mu = file_xattr.get(MU_XATTR)
-        mu = float(mu.decode("UTF8"))
-    except OSError:
-        mu = None
-    try:
-        sigma = file_xattr.get(SIGMA_XATTR)
-        sigma = float(sigma.decode("UTF8"))
-    except OSError:
-        sigma = None
+    mu = get_xattr(filename, MU_XATTR)
+    sigma = get_xattr(filename, SIGMA_XATTR)
+    prev_mu = get_xattr(filename, PREV_MU_XATTR)
+    prev_sigma = get_xattr(filename, PREV_SIGMA_XATTR)
+    prev_index = get_xattr(filename, PREV_INDEX_XATTR)
+    prev_filename = get_xattr(filename, PREV_FILENAME_XATTR)
     previous_rating = False
     if mu is not None or sigma is not None:
         previous_rating = True
-    rating = trueskill.Rating(mu=mu, sigma=sigma)
+        rating = trueskill.Rating(mu=float(mu), sigma=float(sigma))
+    else:
+        rating = trueskill.Rating()
     return {
-        "filename": quote(filename),
+        "filename": filename,
         "mtime": os.path.getmtime(filename),
         "previous_rating": previous_rating,
         "rating": rating,
+        "prev_mu": prev_mu,
+        "prev_sigma": prev_sigma,
+        "prev_index": prev_index,
+        "prev_filename": prev_filename,
         "rank": rank(rating)
     }
 
-def update_file(filename, rating, rm=False):
-    filename = unquote(filename)
-    file_xattr = xattr.xattr(filename)
-    file_xattr.set(MU_XATTR, str(rating.mu).encode('UTF8'))
-    file_xattr.set(SIGMA_XATTR, str(rating.sigma).encode('UTF8'))
+def update_file(file_dict, rating, rm=False):
+    filename = file_dict["filename"]
+    print("Update from %s to %s" % (file_dict["rating"].mu, rating.mu))
+    # Save previous state
+    set_xattr(filename, PREV_MU_XATTR, str(file_dict["rating"].mu))
+    set_xattr(filename, PREV_SIGMA_XATTR, str(file_dict["rating"].sigma))
+    set_xattr(filename, PREV_INDEX_XATTR, str(files_index[filename]))
+    set_xattr(filename, PREV_FILENAME_XATTR, filename)
+    # Update state
+    set_xattr(filename, MU_XATTR, str(rating.mu))
+    set_xattr(filename, SIGMA_XATTR, str(rating.sigma))
     path = pathlib.PurePath(filename)
     suffix = re.split('^(\d+R)\s+?', path.name)[-1]
     new_name = "%sR %s" % (rank(rating), suffix)
@@ -85,14 +108,15 @@ def update_file(filename, rating, rm=False):
     shutil.move(filename, new_filename)
     new_file = get_file(new_filename)
     for i, file in enumerate(eligible_files):
-        if file["filename"] == quote(filename):
+        if file["filename"] == filename:
             if not rm:
                 eligible_files[i] = new_file
             else:
                 del(eligible_files[i])
     return new_file
 
-def update_rankings(win, lose, rm=False):
+def handle_match(win, lose, rm=False):
+    pprint(("handle_match", win, lose, rm))
     try:
         win_file = get_file(win)
         lose_file = get_file(lose)
@@ -108,14 +132,47 @@ def update_rankings(win, lose, rm=False):
     print("After:")
     print("  Win:  " + rank(win_rating) + " " + win_file["filename"])
     print("  Lose: " + rank(lose_rating) + " " + lose_file["filename"])
-    win_file = update_file(win_file["filename"], win_rating)
-    lose_file = update_file(lose_file["filename"], lose_rating, rm)
+    win_file = update_file(win_file, win_rating)
+    lose_file = update_file(lose_file, lose_rating, rm)
     if not rm:
         results = {"win": win_file, "lose": lose_file}
     else:
         results = {"win": win_file, "rm": lose_file}
     pprint(results)
     return results
+
+def undo_update_file(filename, rm=False):
+    # Get previous state
+    prev_mu = get_xattr(filename, PREV_MU_XATTR)
+    prev_sigma = get_xattr(filename, PREV_SIGMA_XATTR)
+    prev_index = get_xattr(filename, PREV_INDEX_XATTR)
+    prev_filename = get_xattr(filename, PREV_FILENAME_XATTR)
+    print("Prev mu: %s" % prev_mu)
+    print("Prev sigma: %s" % prev_sigma)
+    print("Prev filename: %s" % prev_filename)
+    # Restore previous state
+    set_xattr(filename, MU_XATTR, prev_mu)
+    set_xattr(filename, SIGMA_XATTR, prev_sigma)
+    shutil.move(filename, prev_filename)
+    prev_file = get_file(prev_filename)
+    if rm:
+        eligible_files.insert(int(prev_index), prev_file)
+    else:
+        for i, file in enumerate(eligible_files):
+            if file["filename"] == filename:
+                eligible_files[i] = prev_file
+    return prev_filename
+
+def handle_undo(win, lose, rm=False):
+    pprint(("handle_undo", win, lose))
+    win = undo_update_file(win, rm)
+    lose = undo_update_file(lose, rm)
+    if not rm:
+        undo = {"win": get_file(win), "lose": get_file(lose)}
+    else:
+        undo = {"win": get_file(win), "rm": get_file(lose)}
+    pprint(undo)
+    return undo
 
 def load():
     try:
@@ -131,8 +188,11 @@ def load():
         ext = filename.suffix.lower()
         if ext not in SUPPORTED_EXTS:
             continue
+        filename = str(filename)
         # print("Loading: %s" % filename)
-        eligible_files.append(get_file(str(filename)))
+        index = len(eligible_files)
+        files_index[filename] = index
+        eligible_files.append(get_file(filename))
 
 def get_candidates():
     # Select the file with the lowest sigma (confidence) as the starting
@@ -146,7 +206,7 @@ def get_candidates():
             highest_sigma_file = file_dict
             continue
     # Select the file with the closest rank
-    closest_mu_files = []
+    closest_mu_file = None
     lowest_mu_difference = None
     for file_dict in eligible_files:
         # Don't pit an image against itself
@@ -154,36 +214,51 @@ def get_candidates():
             continue
         mu_difference = abs(
             highest_sigma_file['rating'].mu - file_dict['rating'].mu)
-        if len(closest_mu_files) == 0 or mu_difference == lowest_mu_difference:
-            closest_mu_files.append(file_dict)
+        if closest_mu_file is None or mu_difference < lowest_mu_difference:
+            closest_mu_file = file_dict
             lowest_mu_difference = mu_difference
             continue
-        if mu_difference < lowest_mu_difference:
-            closest_mu_files = [file_dict]
-            lowest_mu_difference = mu_difference
-            continue
-    candidates = [highest_sigma_file, random.choice(closest_mu_files)]
+    candidates = [highest_sigma_file, closest_mu_file]
     return candidates
+
+def get_unquote(param):
+    value = request.args.get(param)
+    if value is not None:
+        value = unquote(value)
+    return value
 
 @app.route('/img')
 def img():
-    filename = request.args.get('filename')
+    filename = get_unquote('filename')
     img_file = open(filename, "rb")
     mimetype = mimetypes.guess_type(filename)[0]
     return flask.send_file(img_file, mimetype=mimetype)
 
 @app.route('/')
 def index():
-    win = request.args.get('win')
-    lose = request.args.get('lose')
-    rm = request.args.get('rm')
+    undo = None
     results = None
-    if win is not None:
-        if lose is not None:
-            results = update_rankings(unquote(win), unquote(lose))
-        if rm is not None:
-            results = update_rankings(unquote(win), unquote(rm), rm=True)
+    unwin = get_unquote('unwin')
+    unlose = get_unquote('unlose')
+    unrm = get_unquote('unrm')
+    if unwin is not None:
+        # Undoing something takes preference
+        if unlose is not None:
+            undo = handle_undo(unwin, unlose)
+        if unrm is not None:
+            undo = handle_undo(unwin, unrm, rm=True)
+    else:
+        # Not undoing anything
+        win = get_unquote('win')
+        lose = get_unquote('lose')
+        rm = get_unquote('rm')
+        if win is not None:
+            if lose is not None:
+                results = handle_match(win, lose)
+            if rm is not None:
+                results = handle_match(win, rm, rm=True)
     context = {
+        "undo": undo,
         "results": results,
         "candidates": get_candidates()
     }
